refactor: move setDonatorStatus to services directory
- Updated import paths for setDonatorStatus in the Buy Me A Coffee and Ko-fi webhook routes to point to the new services directory.
- Introduced a new badges service file containing setDonatorStatus and setContentCreatorStatus functions, improving code organization and maintainability.
