+++ src/app/api/webhooks/donation/bmc/route.ts
@@ -1,6 +1,7 @@
 import { headers } from "next/headers";
 import { NextResponse } from "next/server";
 import { db } from "@/db";
+import { setDonatorStatus } from "@/services/badges";
 import {
   getTransactionId,
   verifyBMCWebhook,
@@ -8,7 +9,6 @@ import {
 } from "@/services/donations/bmc";
 import { clerkClient } from "@clerk/nextjs/server";
 
-import { setDonatorStatus } from "@/lib/badges";
 import { donations } from "@/db/schema";
 
 export async function POST(request: Request) {
